requirements.txt

torch>=2.4.0
torchvision>=0.19.0

# --- Hugging Face Ecosystem ---
transformers>=4.44.2   # Recent version required for Florence-2/SigLIP support
datasets>=3.0.0        # For downloading Fashionpedia/DeepFashion
accelerate>=0.34.0     # Optimization for model loading on GPU

# --- Vector Database & Search ---
lancedb>=0.13.0
tantivy>=0.22.0        # CRITICAL: Required for LanceDB Full-Text Search (FTS)
pyarrow>=17.0.0
sentencepiece

# --- Computer Vision Utilities ---
Pillow>=10.4.0         # Image processing
timm>=1.0.9            # PyTorch Image Models (Dependency for Florence-2)
einops>=0.8.0          # Tensor operations (Dependency for Florence-2)

# --- Interactive Dashboard (Jupyter) ---
ipywidgets>=8.1.5
jupyterlab>=4.2.5
tqdm>=4.66.5